refactor(router): split tests into router/tests.rs (<500 lines) + cover fallover timing
